discografia

+++ borrardisco.php
@@ -1,28 +1,32 @@
 <?php
 
-$opc = array(PDO::MYSQL_ATTR_INIT_COMMAND => 'SET NAMES utf8');
-try {
-    $discografia = new PDO('mysql:host=localhost;dbname=discografia', 'discografia', 'discografia', $opc);
-} catch (PDOException $e) {
-    echo 'Falló la conexión: ' . $e->getMessage();
-}
+if (!isset($_SESSION['nombre'])) {
 
 
-$ok = true;
-$discografia->beginTransaction();
+    $opc = array(PDO::MYSQL_ATTR_INIT_COMMAND => 'SET NAMES utf8');
+    try {
+        $discografia = new PDO('mysql:host=localhost;dbname=discografia', 'discografia', 'discografia', $opc);
+    } catch (PDOException $e) {
+        echo 'Falló la conexión: ' . $e->getMessage();
+    }
 
-$codigo = $_GET['codigo'];
-$titulo = $_GET['titulo'];
-$consulta1 = $discografia->query("DELETE FROM cancion WHERE album = '$codigo'");
-$consulta2 = $discografia->query("DELETE FROM album WHERE codigo = '$codigo'");
 
-if ($ok) {
-    $discografia->commit();
-    echo 'Disco ' . $titulo . ' eliminado';
+    $ok = true;
+    $discografia->beginTransaction();
+
+    $codigo = $_GET['codigo'];
+    $titulo = $_GET['titulo'];
+    $consulta1 = $discografia->query("DELETE FROM cancion WHERE album = '$codigo'");
+    $consulta2 = $discografia->query("DELETE FROM album WHERE codigo = '$codigo'");
+
+    if ($ok) {
+        $discografia->commit();
+        echo 'Disco ' . $titulo . ' eliminado';
+    } else {
+        $discografia->rollback();
+    }
+
+    echo '<h4><a href="index.php?">' . "Volver a la página principal" . '</a>';
 } else {
-    $discografia->rollback();
+    header('Location:login.php');
 }
-
-echo '<h4><a href="index.php?">' . "Volver a la página principal" . '</a>';
-
-?>
